Only allow VA/VA tests to alter the ACME challenge ports
