+++ cmd/boulder-va/main.go
@@ -88,8 +88,6 @@ func main() {
 	defer logger.AuditPanic()
 	logger.Info(cmd.VersionString())
 
-	pc := cmd.NewDefaultPortConfig()
-
 	var dnsTimeout time.Duration
 	if c.VA.DNSTimeout != "" {
 		dnsTimeout, err = time.ParseDuration(c.VA.DNSTimeout)
@@ -149,7 +147,6 @@ func main() {
 	}
 
 	vai, err := va.NewValidationAuthorityImpl(
-		pc,
 		resolver,
 		remotes,
 		c.VA.MaxRemoteValidationFailures,
