+++ va/va.go
@@ -20,7 +20,6 @@ import (
 	"github.com/jmhodges/clock"
 	"github.com/letsencrypt/boulder/bdns"
 	"github.com/letsencrypt/boulder/canceled"
-	"github.com/letsencrypt/boulder/cmd"
 	"github.com/letsencrypt/boulder/core"
 	berrors "github.com/letsencrypt/boulder/errors"
 	"github.com/letsencrypt/boulder/features"
@@ -171,6 +170,30 @@ func initMetrics(stats prometheus.Registerer) *vaMetrics {
 	}
 }
 
+// PortConfig specifies what ports the VA should call to on the remote
+// host when performing its checks.
+type PortConfig struct {
+	HTTPPort  int
+	HTTPSPort int
+	TLSPort   int
+}
+
+// NewDefaultPortConfig is a constructor which returns a PortConfig with default
+// settings.
+//
+// CABF BRs section 1.6.1: Authorized Ports: One of the following ports: 80
+// (http), 443 (https), 25 (smtp), 22 (ssh).
+//
+// RFC 8737 section 3: The ACME server initiates a TLS connection to the chosen
+// IP address. This connection MUST use TCP port 443.
+func NewDefaultPortConfig() *PortConfig {
+	return &PortConfig{
+		HTTPPort:  80,
+		HTTPSPort: 443,
+		TLSPort:   443,
+	}
+}
+
 // ValidationAuthorityImpl represents a VA
 type ValidationAuthorityImpl struct {
 	vapb.UnimplementedVAServer
@@ -193,7 +216,6 @@ type ValidationAuthorityImpl struct {
 
 // NewValidationAuthorityImpl constructs a new VA
 func NewValidationAuthorityImpl(
-	pc *cmd.PortConfig,
 	resolver bdns.Client,
 	remoteVAs []RemoteVA,
 	maxRemoteFailures int,
@@ -208,6 +230,8 @@ func NewValidationAuthorityImpl(
 	if features.Enabled(features.CAAAccountURI) && len(accountURIPrefixes) == 0 {
 		return nil, errors.New("no account URI prefixes configured")
 	}
+
+	pc := NewDefaultPortConfig()
 
 	va := &ValidationAuthorityImpl{
 		log:                logger,
