+++ va/va_test.go
@@ -19,7 +19,6 @@ import (
 
 	"github.com/jmhodges/clock"
 	"github.com/letsencrypt/boulder/bdns"
-	"github.com/letsencrypt/boulder/cmd"
 	"github.com/letsencrypt/boulder/core"
 	corepb "github.com/letsencrypt/boulder/core/proto"
 	"github.com/letsencrypt/boulder/features"
@@ -123,15 +122,7 @@ func setup(srv *httptest.Server, maxRemoteFailures int, userAgent string, remote
 		userAgent = "user agent 1.0"
 	}
 
-	portConfig := cmd.NewDefaultPortConfig()
-	if srv != nil {
-		port := getPort(srv)
-		portConfig.HTTPPort = port
-		portConfig.TLSPort = port
-	}
 	va, err := NewValidationAuthorityImpl(
-		// Use the test server's port as both the HTTPPort and the TLSPort for the VA
-		portConfig,
 		&bdns.MockClient{Log: logger},
 		nil,
 		maxRemoteFailures,
@@ -142,6 +133,15 @@ func setup(srv *httptest.Server, maxRemoteFailures int, userAgent string, remote
 		logger,
 		accountURIPrefixes,
 	)
+
+	// Adjusting industry regulated ACME challenge port settings is fine during
+	// testing
+	if srv != nil {
+		port := getPort(srv)
+		va.httpPort = port
+		va.tlsPort = port
+	}
+
 	if err != nil {
 		panic(fmt.Sprintf("Failed to create validation authority: %v", err))
 	}
